+++ packages/react-native-sdk/src/index.ts
@@ -8,6 +8,7 @@ import 'intl-pluralrules';
 import { registerGlobals } from '@stream-io/react-native-webrtc';
 import Logger from '@stream-io/react-native-webrtc/src/Logger';
 import { Platform } from 'react-native';
+import { registerSDKGlobals } from './utils/internal/registerSDKGlobals';
 import './utils/keepCallAliveHeadlessTask';
 
 // We're registering globals, because our video JS client is serving SDKs that use browser based webRTC functions.
@@ -15,6 +16,7 @@ import './utils/keepCallAliveHeadlessTask';
 // Reference: https://github.com/react-native-webrtc/react-native-webrtc/blob/16cff1523da457dbcc27bb0744ee2bad3a987c41/Documentation/BasicUsage.md#registering-globals
 if (Platform.OS !== 'web') {
   registerGlobals();
+  registerSDKGlobals();
 }
 
 // only enable warning and error logs from webrtc library