+++ sample-apps/react-native/dogfood/ios/AppDelegate.swift
@@ -91,8 +91,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
     
     let options = WebRTCModuleOptions.sharedInstance()
     options.enableMultitaskingCameraAccess = true
-    //  uncomment below to see native webrtc logs
-    //  options.loggingSeverity = .info // Use enum value directly
+    #if DEBUG
+    // Native WebRTC logs (debug-only)
+    options.loggingSeverity = .verbose
+    #endif
 
     let delegate = ReactNativeDelegate()
     let factory = RCTReactNativeFactory(delegate: delegate)