+++ sample-apps/react-native/dogfood/src/screens/LiveStream/ViewLiveStream.tsx
@@ -43,7 +43,7 @@ export const ViewLiveStreamChildren = ({
   } = route;
 
   useEffect(() => {
-    callManager.start({ audioRole: 'listener' });
+    callManager.start({ audioRole: 'listener', enableStereoAudioOutput: true });
     return () => {
       callManager.stop();
     };